Wizard-Klassen überarbeiten
* Klassenstruktur vereinfachen
* unnötige Klassen entfernen
* Allgemeine Wizardklassen in von Workingtime in allgemeines Projekt
verschieben
