+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/widgets/MinovaWizard.java
@@ -1,4 +1,4 @@
-package aero.minova.workingtime.wizard;
+package aero.minova.rcp.rcp.widgets;
 
 import javax.inject.Inject;
 
@@ -7,8 +7,9 @@ import org.eclipse.e4.core.contexts.IEclipseContext;
 import org.eclipse.jface.dialogs.DialogSettings;
 import org.eclipse.jface.dialogs.IDialogSettings;
 import org.eclipse.jface.wizard.IWizardPage;
+import org.eclipse.jface.wizard.Wizard;
 
-import aero.minova.rcp.rcp.widgets.AbstractWizard;
+import aero.minova.rcp.model.form.MDetail;
 
 /**
  * Basisklasse für alle Minova-Assistenten (Wizards)
@@ -16,12 +17,14 @@ import aero.minova.rcp.rcp.widgets.AbstractWizard;
  * @author erlanger
  * @since 12.0.0
  */
-public class MinovaWizard extends AbstractWizard {
-	private IWizardFinishAction finishAction;
+public class MinovaWizard extends Wizard {
+	private IMinovaWizardFinishAction finishAction;
 	private Object opener;
 
 	@Inject
 	private IEclipseContext context;
+
+	protected MDetail originalMDetail;
 
 	public MinovaWizard(String wizardName) {
 		this.setWindowTitle(wizardName);
@@ -50,7 +53,7 @@ public class MinovaWizard extends AbstractWizard {
 	/**
 	 * @return the finishAction
 	 */
-	public IWizardFinishAction getFinishAction() {
+	public IMinovaWizardFinishAction getFinishAction() {
 		return finishAction;
 	}
 
@@ -94,7 +97,7 @@ public class MinovaWizard extends AbstractWizard {
 	 * @param finishAction
 	 *            the finishAction to set
 	 */
-	public void setFinishAction(IWizardFinishAction finishAction) {
+	public void setFinishAction(IMinovaWizardFinishAction finishAction) {
 		this.finishAction = finishAction;
 	}
 
@@ -106,5 +109,9 @@ public class MinovaWizard extends AbstractWizard {
 	 */
 	public void setOpener(Object opener) {
 		this.opener = opener;
+	}
+
+	public void setOriginalMDetail(MDetail originalMDetail) {
+		this.originalMDetail = originalMDetail;
 	}
 }
