+++ bundles/aero.minova.workingtime.wizard/src/aero/minova/workingtime/wizard/FillWorkingtimeWizard.java
@@ -20,6 +20,7 @@ import aero.minova.rcp.dialogs.NotificationPopUp;
 import aero.minova.rcp.model.SqlProcedureResult;
 import aero.minova.rcp.model.Table;
 import aero.minova.rcp.preferences.ApplicationPreferences;
+import aero.minova.rcp.rcp.widgets.MinovaWizard;
 
 /**
  * Mit diesem Assistent (Wizard) kann man Arbeitszeiten über mehrere Tage auffüllen
