+++ bundles/aero.minova.workingtime.wizard/src/aero/minova/workingtime/wizard/FinishAction.java
@@ -1,9 +1,12 @@
 package aero.minova.workingtime.wizard;
 
+import aero.minova.rcp.rcp.widgets.IMinovaWizardFinishAction;
+import aero.minova.rcp.rcp.widgets.MinovaWizard;
+
 /**
  * Arbeitszeit auffüllen mit den im Wizard angegebenen Daten
  */
-public class FinishAction implements IWizardFinishAction {
+public class FinishAction implements IMinovaWizardFinishAction {
 	private MinovaWizard wizard;
 
 	@Override
